+++ Gruntfile.js
@@ -19,12 +19,12 @@ module.exports = function (grunt)
             dist: {
                 files: {
                     '<%= pkg.folders.dist %>/jquery.steps.js': [
+                        '<%= pkg.folders.src %>/helper.js',
                         '<%= pkg.folders.src %>/privates.js',
                         '<%= pkg.folders.src %>/publics.js',
                         '<%= pkg.folders.src %>/enums.js',
                         '<%= pkg.folders.src %>/model.js',
-                        '<%= pkg.folders.src %>/defaults.js',
-                        '<%= pkg.folders.src %>/helper.js'
+                        '<%= pkg.folders.src %>/defaults.js'
                     ]
                 }
             }